logger.tsの中で行っていたテストを分割
